mover archivos de prueba un directorio para arriba
