+++ tests/pytest/archivos_de_prueba/clases_nominal.xlsx
